Progress on Arcane Archer
added a new function to check if the weapon being used is part of a passed in list of allowed weapons, this is in new mechanics and is functionally almost the same as the check if feature present ability.

added logic for spellcasting, added logic for arrow enhancement. Currently they're just always on/off toggles, but need to be changed to be toggle an option and it remains on till the following morning when another can be toggled.
The flaming, shocking, frost need to be mutually exclusive of the burst versions. The ability needs the temporary enchantment bonus totally removed and it needs to allow the aligned in addition to the elemental options.

Need a replacement option for distance as well.

Ability is more or less a copy-paste of the warpriest enhancements at the moment, so it needs work to remove the uneeded options.

Still, progress